test(duellingpicklist): add accessibility tests
